+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 ### [@project-version@](https://github.com/wow-addon-dev/Percursus/tree/@project-version@) (@build-date@)
 [Full Changelog](@full-changelog@) - [Previous Releases](https://github.com/wow-addon-dev/Percursus/releases)
 
-- Minor code adjustments
+- Added: Race times for the event 'Northrend Cup'
+- Removed: TOC version for patch 12.0.7 [retail]